correct enable was not unsigned

+++ projects/vm/incs/socket_inc.h
@@ -6,23 +6,23 @@
 /*   By: fbenneto <fbenneto@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/03/28 15:50:14 by fbenneto          #+#    #+#             */
-/*   Updated: 2019/03/29 09:26:04 by fbenneto         ###   ########.fr       */
+/*   Updated: 2019/03/29 10:20:31 by fbenneto         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef SOCKET_INC_H
 #define SOCKET_INC_H
 
+#include <arpa/inet.h>
 #include <netinet/in.h>
 #include <sys/socket.h>
-#include <arpa/inet.h>
 #include <unistd.h>
 
 #define MAX_CLIENT_SOCKET 1
 #define SOCK_SEND_FLAGS 0
 
-# define ENABLE_SOCKET 1
-# define DISABLE_SOCKET 0
+#define ENABLE_SOCKET 1
+#define DISABLE_SOCKET 0
 #define ALLOW_REUSE_ADDR 1
 
 typedef struct sockaddr_in t_sockaddr_in;
@@ -38,11 +38,11 @@ typedef struct s_socket
 {
 	char *		 ip;
 	int16_t		 port;
-	int			 enable : 1;
+	uint8_t		 enable : 1;
 	t_sock_inter server;
 
 	t_sock_inter clients[MAX_CLIENT_SOCKET];
-	int			 nb_client;
+	uint16_t	 nb_client;
 } t_socket;
 
 #endif // !SOCKET_INC_H
